moving things to right places
